+++ uplink/metainfo/kvmetainfo/project.go
@@ -4,9 +4,9 @@
 package kvmetainfo
 
 import (
-	"storj.io/storj/pkg/storage/buckets"
 	"storj.io/storj/uplink/eestream"
 	"storj.io/storj/uplink/metainfo"
+	"storj.io/storj/uplink/storage/buckets"
 	"storj.io/storj/uplink/storage/streams"
 )
